#106259

Related work items: #106259, #106399

+++ Logitude/Logitude.BL/CommonDataModel/Tools/EntityService/DocumentsFilingService.cs
@@ -1856,23 +1856,35 @@ namespace Logitude.BL.CommonDataModel.Tools.EntityService
 
         private void CreateCustomsDocumentsFilingMetaDataValue(DocumentsFilingMetaDataValuePM itemPM)
         {
-            CustomsDocumentMetaDataValue customsDocumentMetaDataValue = new CustomsDocumentMetaDataValue();
-            customsDocumentMetaDataValue.MetaDataTypeCode = itemPM.DocumentsMetaDataTypeCode;
-            customsDocumentMetaDataValue.CustomsDocumentId = entityPM.Id;
-            customsDocumentMetaDataValue.Tenant = tenant; 
-            customsDocumentMetaDataValue.MetaDataValue = itemPM.MetaDataValue;
-            customsDocumentMetaDataValueRepository.Add(customsDocumentMetaDataValue);
+            CustomsDocumentQueryService myCustomsDocumentQueryService = new CustomsDocumentQueryService(tenant);
+            string documentsFilingId = myCustomsDocumentQueryService.GetDocumentInIdByCustomsDocId(itemPM.Id, tenant);
+
+            if (documentsFilingId != null)
+            {
+                CustomsDocumentMetaDataValue customsDocumentMetaDataValue = new CustomsDocumentMetaDataValue();
+                customsDocumentMetaDataValue.MetaDataTypeCode = itemPM.DocumentsMetaDataTypeCode;
+                customsDocumentMetaDataValue.CustomsDocumentId = entityPM.Id;
+                customsDocumentMetaDataValue.Tenant = tenant;
+                customsDocumentMetaDataValue.MetaDataValue = itemPM.MetaDataValue;
+                customsDocumentMetaDataValueRepository.Add(customsDocumentMetaDataValue);
+            }
         }
         private void UpdateCustomsDocumentsFilingMetaDataValue(DocumentsFilingMetaDataValuePM itemPM)
         {
+
             CustomsDocumentMetaDataValue customsDocumentMetaDataValue = customsDocumentMetaDataValueRepository.GetCustomsDocumentMetaDataValuesByCustomDocumentAndMetaDateValue(entityPM.Id, tenant, itemPM.DocumentsMetaDataTypeCode);
-            customsDocumentMetaDataValue.MetaDataValue=itemPM.MetaDataValue;
-            customsDocumentMetaDataValueRepository.Update(customsDocumentMetaDataValue);
+            if(customsDocumentMetaDataValue != null) {
+                customsDocumentMetaDataValue.MetaDataValue = itemPM.MetaDataValue;
+                customsDocumentMetaDataValueRepository.Update(customsDocumentMetaDataValue);
+            }
+           
         }
         private void DeleteCustomsDocumentsFilingMetaDataValue(DocumentsFilingMetaDataValuePM itemPM)
         {
-            CustomsDocumentMetaDataValue customsDocumentMetaDataValue = customsDocumentMetaDataValueRepository.GetCustomsDocumentMetaDataValuesByCustomDocumentAndMetaDateValue(entityPM.Id, tenant, itemPM.DocumentsMetaDataTypeCode);
-            customsDocumentMetaDataValueRepository.Remove(customsDocumentMetaDataValue);
+
+            CustomsDocumentMetaDataValue customsDocumentMetaDataValue = customsDocumentMetaDataValueRepository.GetCustomsDocumentMetaDataValuesByCustomDocumentAndMetaDateValue(entityPM.Id, tenant,itemPM.DocumentsMetaDataTypeCode);
+            if(customsDocumentMetaDataValue != null) 
+               customsDocumentMetaDataValueRepository.Remove(customsDocumentMetaDataValue);
         }
 
         public string RandomString(int length)
